2017-04-21 13:22 commit

+++ CPSS/Data/CPSS.Data.DataAccess.Interfaces/Basic/Parameters/QuerySubCompanyListParameter.cs
@@ -25,7 +25,7 @@ namespace CPSS.Data.DataAccess.Interfaces.Basic.Parameters
 
         public string LinkTel { set; get; }
 
-        public int Status { set; get; }
+        public short Status { set; get; }
 
         public short Deleted { get; set; }
     }
